Updated ICE Import logic to consume configuration JSON files

+++ ICE_Import/ICE_Import/PushTable.cs
@@ -98,7 +98,7 @@ namespace ICE_Import
                 {
                     DateTime expirationDate = TMLDBReader.GetExpirationDate(
                         "future",
-                        IdInstrument,
+                        (long)IdInstrument,
                         future.StripName,
                         ref log);
 
@@ -193,7 +193,7 @@ namespace ICE_Import
                         (double)option.EOD_Option.StrikePrice.GetValueOrDefault(),
                         (double)OptionStrikeIncrement,
                         (double)OptionStrikeDisplay,
-                        IdInstrument);
+                        (long)IdInstrument);
 
                     newOption = !OptionNameHashSet.Contains(optionName);
 
@@ -201,7 +201,7 @@ namespace ICE_Import
                     {
                         DateTime expirationDate = TMLDBReader.GetExpirationDate(
                             "option",
-                            IdInstrument,
+                            (long)IdInstrument,
                             option.EOD_Option.StripName,
                             ref log);
 
@@ -265,7 +265,7 @@ namespace ICE_Import
                         0.5,
                         riskFreeInterestRate,
                         Utilities.NormalizePrice(option.EOD_Option.SettlementPrice),
-                        TickSize);
+                        (double)TickSize);
 
                         if (object.ReferenceEquals(impliedvol, null) || double.IsNaN(impliedvol) || double.IsInfinity(impliedvol))
                         {
